requirements.txt
Flask>=2.2
python-dotenv>=0.21
pymongo>=4.0
flask-login>=0.6
Flask-WTF>=1.0
openpyxl>=3.0
email-validator>=1.3
gunicorn>=20.0
pytest>=7.0
mongomock>=4.0
